+++ addon/utils/chart-types.js
@@ -1,8 +1,0 @@
-/* Map ember-highcharts modes to Highcharts methods
- * https://api.highcharts.com/class-reference/Highcharts.html
- */
-export const CHART_TYPES = {
-  StockChart: 'stockChart',
-  Map: 'mapChart',
-  undefined: 'chart',
-};