+++ tests/dummy/app/components/chart-highstock-interactive.js
@@ -1,31 +1,34 @@
-import { copy } from 'ember-copy';
+import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
-import Component from '@ember/component';
+import { action } from '@ember/object';
+
+import { copy } from 'ember-copy';
 import stockData from '../data/stock';
 
-export default Component.extend({
-  dynamicChart: service('dynamic-chart'),
+export default class BarBasic extends Component {
+  @service('dynamic-chart')
+  dynamicChart;
 
-  chartOptions: {
+  chartOptions = {
     rangeSelector: {
       selected: 1
     },
     title: {
       text: 'Highstock: AAPL Stock Price'
     }
-  },
+  };
 
-  chartData: copy(stockData, true),
+  chartData = copy(stockData, true);
 
-  actions: {
-    updateSeriesData() {
-      let newChartData = this.get('dynamicChart').updateSeriesData(stockData, 100, 514);
-      this.set('chartData', newChartData);
-    },
-
-    setSeriesCount(numSeries) {
-      let newChartData = this.get('dynamicChart').updateSeriesCount(stockData, numSeries);
-      this.set('chartData', newChartData);
-    }
+  @action
+  updateSeriesData() {
+    let newChartData = this.dynamicChart.updateSeriesData(stockData, 100, 514);
+    this.chartData = newChartData;
   }
-});
+
+  @action
+  setSeriesCount(numSeries) {
+    let newChartData = this.dynamicChart.updateSeriesCount(stockData, numSeries);
+    this.chartData = newChartData;
+  }
+}