+++ tests/dummy/app/components/chart-line-interactive.js
@@ -1,12 +1,14 @@
 import { copy } from 'ember-copy';
 import { inject as service } from '@ember/service';
-import Component from '@ember/component';
+import Component from '@glimmer/component';
 import commitStats from '../data/commit-stats';
+import { action } from '@ember/object';
 
-export default Component.extend({
-  dynamicChart: service('dynamic-chart'),
+export default class BarBasic extends Component {
+  @service('dynamic-chart')
+  dynamicChart;
 
-  chartOptions: {
+  chartOptions = {
     chart: {
       type: 'line'
     },
@@ -24,30 +26,32 @@ export default Component.extend({
         text: '# of Commits'
       }
     }
-  },
-
-  chartData: copy(commitStats, true),
-
-  actions: {
-    updateSeriesData() {
-      let newChartData = this.get('dynamicChart').updateSeriesData(commitStats, 2, 52);
-      this.set('chartData', newChartData);
-    },
-
-    fullUpdateToSeries() {
-      let newChartData = this.get('dynamicChart').updateSeriesData(commitStats, 2, 52);
-
-      // updated currentTime attribute causes series.update() to be used instead of series.setData()
-      newChartData.forEach((series) => {
-        series.currentTime = Date.now();
-      });
-
-      this.set('chartData', newChartData);
-    },
-
-    setSeriesCount(numSeries) {
-      let newChartData = this.get('dynamicChart').updateSeriesCount(commitStats, numSeries);
-      this.set('chartData', newChartData);
-    }
   }
-});
+
+  chartData = copy(commitStats, true);
+
+  @action
+  updateSeriesData() {
+    let newChartData = this.dynamicChart.updateSeriesData(commitStats, 2, 52);
+    this.chartData = newChartData;
+  }
+
+  @action
+  fullUpdateToSeries() {
+    let newChartData = this.dynamicChart.updateSeriesData(commitStats, 2, 52);
+
+    // updated currentTime attribute causes series.update() to be used instead of series.setData()
+    newChartData.forEach((series) => {
+      series.currentTime = Date.now();
+    });
+
+    this.chartData = newChartData;
+  }
+
+  @action
+  setSeriesCount(numSeries) {
+    let newChartData = this.dynamicChart.updateSeriesCount(commitStats, numSeries);
+    this.chartData = newChartData;
+  }
+
+}